+++ .github/Dockerfile.dev
@@ -1,4 +1,4 @@
-FROM python:3.12
+FROM python:3.11
 
 # Set the working directory to /app
 WORKDIR /app
