+++ compose.sample.yml
@@ -5,7 +5,7 @@ services:
     volumes:
       - YOUR_OUTPUT_DIR:/output
     environment:
-      - OMNIVORE_API_KEY=YOUR_API_KEY
-      - OPENAI_API_KEY=YOUR_API_KEY
-      - ANTHROPIC_API_KEY=YOUR_API_KEY
+      - OMNIVORE_API_KEY=${OMNIVORE_API_KEY}
+      - OPENAI_API_KEY=${OPENAI_API_KEY}
+      - ANTHROPIC_API_KEY=${ANTHROPIC_API_KEY}
     restart: unless-stopped
