+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 version = "0.22.1"
-requires-python = ">=3.12"
+requires-python = ">=3.11"
 name = "memorymarker"
 description = "memorymarker"
 dependencies = [
@@ -8,7 +8,7 @@ dependencies = [
     "instructor>=0.6.2",
     "iterpy>=1.6.0",
     "joblib>=1.3.2",
-    "omnivoreql>=0.2.1",
+    "omnivoreql>=0.3.3",
     "openai>=1.13.3",
     "pydantic>=2.6.2",
     "python-dotenv>=1.0.1",
@@ -19,7 +19,7 @@ dependencies = [
     "anthropic>=0.21.3",
 ]
 authors = [{ name = "Martin Bernstorff", email = "martinbernstorff@gmail.com" }]
-classifiers = ["Programming Language :: Python :: 3.12"]
+classifiers = ["Programming Language :: Python :: 3.11"]
 
 [project.license]
 file = "LICENSE"
@@ -31,7 +31,7 @@ content-type = "text/markdown"
 [tool]
 rye = { dev-dependencies = [
     "diff-cover==8.0.3",
-    "pyright==1.1.350",
+    "pyright>=1.1.368",
     "pytest>=7.4.0",
     "pytest-cov==4.1.0",
     "pytest-xdist==3.5.0",
