+++ requirements.lock
@@ -18,6 +18,7 @@ anthropic==0.21.3
     # via memorymarker
 anyio==3.7.1
     # via anthropic
+    # via gql
     # via httpx
     # via openai
 attrs==23.2.0
@@ -48,7 +49,7 @@ frozenlist==1.4.1
     # via aiosignal
 fsspec==2024.3.1
     # via huggingface-hub
-gql==3.4.1
+gql==3.5.0
     # via omnivoreql
 graphql-core==3.2.3
     # via gql
@@ -81,7 +82,7 @@ mdurl==0.1.2
 multidict==6.0.5
     # via aiohttp
     # via yarl
-omnivoreql==0.2.1
+omnivoreql==0.3.3
     # via memorymarker
 openai==1.13.3
     # via instructor
@@ -101,6 +102,7 @@ pygments==2.17.2
     # via rich
 python-dotenv==1.0.1
     # via memorymarker
+    # via omnivoreql
 pytz==2024.1
     # via memorymarker
 pyyaml==6.0.1
